Run commands and conversion to BA format
Through Class Command executing of commands was made possible - Finding the smallest cut and drawing it in JavaFX.
Also conversion from Graph Model to BA format was added to enable using ba-graph library.

+++ src/romulo/GraphView.java
@@ -11,7 +11,7 @@ public class GraphView extends javafx.scene.layout.Pane {
         edges.add(e);
         this.getChildren().add(e);
     }
-    void addVertex(int x, int y, int r) {
+    void addVertex(double x, double y, int r) {
         VertexView v = new VertexView(x, y, r, vertices.size());
         vertices.add(v);
         this.getChildren().add(v);
